requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==0.27.8
requests==2.31.0
pydantic==2.0.0
pytest==7.4.0
